+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v2.11.1] - 2020-07-03
+## Fixed
+- Mark snapshot as sent only if sent to at least one provider
+
 ## [v2.11.0] - 2020-07-02
 ## Changed
 - Revert: remove stale tips at higher height